# gabarita-ai-backend/docs/swagger.yaml
openapi: 3.0.3
info:
  title: Gabarita AI API
  description: |
    API do sistema Gabarita AI para geração de questões, gerenciamento de usuários e planos.
    
    ## Autenticação
    A API utiliza autenticação via JWT tokens. Inclua o token no header:
    ```
    Authorization: Bearer <seu_token_jwt>
    ```
    
    ## Formato de Resposta
    Todas as respostas seguem o padrão:
    ```json
    {
      "success": true|false,
      "data": { /* dados da resposta */ },
      "message": "Mensagem descritiva",
      "error": "Tipo do erro (apenas em caso de erro)"
    }
    ```
  version: 1.0.0
  contact:
    name: Gabarita AI Support
    email: suporte@gabaritai.app.br
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://gabarita-ai-backend.onrender.com
    description: Servidor de Produção
  - url: http://localhost:5000
    description: Servidor de Desenvolvimento
paths:
...
  /api/auth/login:
    post:
      tags:
        - Autenticação
      summary: Login do usuário
      description: Autentica usuário e retorna token JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: usuario@exemplo.com
                password:
                  type: string
                  format: password
                  example: minhasenha123
      responses:
        '200':
          description: Login realizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/auth/logout:
    post:
      tags:
        - Autenticação
      summary: Logout do usuário
      description: Invalida o token JWT atual
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout realizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
...
  /api/login:
    post:
      tags:
        - Autenticação (Aliases)
      summary: Login (Alias)
      description: Alias para /api/auth/login
      requestBody:
        $ref: '#/paths/~1api~1auth~1login/post/requestBody'
      responses:
        $ref: '#/paths/~1api~1auth~1login/post/responses'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Operação realizada com sucesso
      required:
        - success
        - message
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
          example: Erro de validação
        message:
          type: string
          example: Descrição detalhada do erro
      required:
        - success
        - error
        - message
    AuthResponse:
      allOf:
        - $ref: '#/components/schemas/SuccessResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                token:
                  type: string
                  description: JWT token
                user:
                  $ref: '#/components/schemas/User'
              required:
                - token
                - user
    User:
      type: object
      properties:
        id:
          type: integer
          description: ID único do usuário
        email:
          type: string
          format: email
          description: Email do usuário
        name:
          type: string
          description: Nome completo do usuário
        created_at:
          type: string
          format: date-time
          description: Data de criação da conta
        updated_at:
          type: string
          format: date-time
          description: Data da última atualização
        plan:
          type: string
          nullable: true
          description: Plano atual do usuário
      required:
        - id
        - email
        - name
        - created_at
        - updated_at
...